Moved bucket components to their own subdirectory.
