+++ scripts/components/bucket/BucketTabs.js
