+++ scripts/containers/BucketGroupsPage.js
@@ -1,6 +1,6 @@
 import { bindActionCreators } from "redux";
 import { connect } from "react-redux";
-import BucketGroups from "../components/BucketGroups";
+import BucketGroups from "../components/bucket/BucketGroups";
 import * as BucketActions from "../actions/bucket";
 import * as NotificationsActions from "../actions/notifications";
 import { push as updatePath } from "react-router-redux";
